feat: split Props in a different file
